test: extract zoe-startInstance test file
